Add support for a CredentialsProvider in GitConfigurationSource
Summary:
Add support for a CredentialsProvider in GitConfigurationSource
The plan is to upload this jar to Signifyd Nexus so we can use it for config files in our private github.
I will also try to submit a pull request.

Test Plan: Added unit test. Also installed the jar locally and tested against a private repo.

Reviewers: nikhil

Reviewed By: nikhil

JIRA Issues: RAP-1030

Differential Revision: https://signifyd.phacility.com/D7770